model_tree: iter.rs, tests.rs
